+++ quint/src/graphics.ts
@@ -10,53 +10,84 @@
 
 import chalk from 'chalk'
 import { strict as assert } from 'assert'
-import { range } from 'lodash'
+import {
+  Doc, braces, brackets, docJoin, format, group, line, linebreak,
+  nest, parens, richtext, text, textify
+} from './prettierimp'
 
 import { QuintEx } from './quintIr'
 import { ExecutionFrame } from './runtime/trace'
 import { zerog } from './idGenerator'
 
-// convert a Quint expression to a colored string, tuned for REPL
-export function chalkQuintEx(ex: QuintEx): string {
+/**
+ * An abstraction of a Console of bounded text width.
+ */
+export interface Console {
+  width: number,
+  out: (s: string) => void
+}
+
+/**
+ * Find out the terminal width for text formatting.
+ * Since this number may change while running, it is a function.
+ * Also, when the output is redirected, the number of columns is undefined.
+ */
+export const terminalWidth = () => {
+  const cols = process.stdout.columns
+  return (cols !== undefined && cols > 0) ? cols : 80
+}
+
+// convert a Quint expression to a colored pretty-printed doc, tuned for REPL
+export function prettyQuintEx(ex: QuintEx): Doc {
+  // a helper function to produce specific indentation
+  const nary = (left: Doc, args: Doc[], right: Doc, padding: Doc = linebreak): Doc => {
+    const as = group([
+      nest('  ', [ padding, docJoin([ text(','), line() ], args) ]),
+      padding
+    ])
+    return group([left, as, right])
+  }
+
   switch (ex.kind) {
     case 'bool':
-      return chalk.yellow(`${ex.value}`)
+      return richtext(chalk.yellow, ex.value.toString())
 
     case 'int':
-      return chalk.yellow(`${ex.value}`)
+      return richtext(chalk.yellow, ex.value.toString())
 
     case 'str':
-      return chalk.green(`"${ex.value}"`)
+      return richtext(chalk.green, `"${ex.value}"`)
 
     case 'app':
       switch (ex.opcode) {
-        case 'Set': {
-          const as = ex.args.map(chalkQuintEx).join(', ')
-          return chalk.green('Set') + `(${as})`
-        }
+        case 'Set':
+          return group([
+            richtext(chalk.green, 'Set'),
+            nary(text('('), ex.args.map(prettyQuintEx), text(')'))
+          ])
 
         case 'Map': {
-          const ps = ex.args.map(tup => {
+          const ps: Doc[] = ex.args.map(tup => {
             if (tup.kind === 'app' &&
                    tup.opcode === 'Tup' && tup.args.length === 2) {
               const [k, v] = tup.args
-              return `${chalkQuintEx(k)} -> ${chalkQuintEx(v)}`
+              return group([
+                prettyQuintEx(k),
+                richtext(chalk.gray, ' ->'),
+                nest('  ', [line(), prettyQuintEx(v)])
+              ])
             } else {
-              return '<expected-pair>'
+              return text('<expected-pair>')
             }
           })
-          const as = ps.join(', ')
-          return chalk.green('Map') + `(${as})`
+          return group([richtext(chalk.green, 'Map'), nary(text('('), ps, text(')'))])
         }
 
-        case 'Tup': {
-          const as = ex.args.map(chalkQuintEx).join(', ')
-          return `(${as})`
-        }
+        case 'Tup':
+          return nary(text('('), ex.args.map(prettyQuintEx), text(')'))
 
         case 'List': {
-          const as = ex.args.map(chalkQuintEx).join(', ')
-          return `[${as}]`
+          return nary(text('['), ex.args.map(prettyQuintEx), text(']'))
         }
 
         case 'Rec': {
@@ -64,85 +95,113 @@ export function chalkQuintEx(ex: QuintEx): string {
           for (let i = 0; i < ex.args.length / 2; i++) {
             const key = ex.args[2 * i]
             if (key && key.kind === 'str') {
-              const value = chalkQuintEx(ex.args[2 * i + 1])
-              kvs.push(`${chalk.green(key.value)}: ${value}`)
+              const value = prettyQuintEx(ex.args[2 * i + 1])
+              kvs.push(group([
+                text(key.value),
+                richtext(chalk.gray, ':'),
+                nest('  ', [line(), value]),
+              ]))
             }
           }
-          return `{ ${kvs.join(', ')} }`
+          return nary(text('{'), kvs, text('}'), line())
         }
 
         default:
           // instead of throwing, show it in red
-          return chalk.red(`unsupported operator: ${ex.opcode}(...)`)
+          return richtext(chalk.red, `unsupported operator: ${ex.opcode}(...)`)
       }
 
     case 'lambda': {
-      const params = ex.params.map(p => p.name).join(', ')
-      return `{ (${params}) => ... }`
+      const params = 
+        parens(docJoin([text(','), line()], ex.params.map(p => text(p.name))))
+
+      return braces(group(textify([ params, line(), '=>', line(), '...' ])))
     }
 
     default:
-      return chalk.red(`unsupported operator: ${ex.kind}`)
+      return richtext(chalk.red, `unsupported operator: ${ex.kind}`)
   }
 }
 
+/**
+ * Print an execution frame and its children recursively.
+ * Since this function is printing a tree, we need precise text alignment.
+ * Using a tree here with an optional line break would produce incorrect output.
+ * 
+ * @param console the console to print in
+ * @param frame the frame to print
+ * @param isLast the array of booleans, one per ancestor, that indicates whether
+ *       an ancestor does not have siblings to the right, the last index
+ *       corresponds to the direct parent.
+ */
 export function
-printExecutionFrameRec(out: (line: string) => void,
-    frame: ExecutionFrame,
-    isLast: boolean[]) {
+printExecutionFrameRec(console: Console,
+    frame: ExecutionFrame, isLast: boolean[]): void {
   // convert the arguments and the result to strings
-  const args =
-    frame.args.map(a => chalkQuintEx(a.toQuintEx(zerog))).join(', ')
+  const args = docJoin(
+    [ text(','), line() ],
+    frame.args.map(a => prettyQuintEx(a.toQuintEx(zerog)))
+  )
   const r =
     frame.result.isNone()
-      ? 'none'
-      : chalkQuintEx(frame.result.value.toQuintEx(zerog))
+      ? text('none')
+      : prettyQuintEx(frame.result.value.toQuintEx(zerog))
   const depth = isLast.length
   // generate the tree ASCII graphics for this frame
   let treeArt = isLast.map((il, i) =>
     (i < depth - 1)
       // continue the ancestor's branch, unless it's the last one
-      ? (il ? '   ' : '│  ')
+      ? il ? '   ' : '│  '
       // close or close & continue the leaf branch,
       // depending on whether this frame is the last one
-      : (il ? '└─ ' : '├─ ')
+      : il ? '└─ ' : '├─ '
   ).join('')
-  out(`${treeArt}${frame.app.opcode}(${args}) => ${r}`)
+  console.out(treeArt)
+
+  // format the call with its arguments and place it right after the tree art
+  const argsDoc =
+    group(textify(['(', nest('  ', [linebreak, group(args)]), linebreak, ')']))
+  // draw proper branches in the indentation
+  const indentTreeArt = isLast.map(il => il ? '   ' : '│  ').join('')
+  // pretty print the arguments and the result
+  const callDoc = group(
+    nest(indentTreeArt, [
+      text(frame.app.opcode),
+      group([ argsDoc, nest('  ', group([text(' =>'), line(), r])) ]),
+    ]))
+
+  console.out(format(console.width, treeArt.length, callDoc))
+  console.out('\n')
   const n = frame.subframes.length
   // visualize the children
   frame.subframes.forEach((f, i) =>
-    printExecutionFrameRec(out, f, isLast.concat([i === n - 1]))
+    printExecutionFrameRec(console, f, isLast.concat([i === n - 1]))
   )
 }
 
 /**
  * Print a trace with chalk.
  */
-export function printTrace(out: (line: string) => void,
-                           states: QuintEx[],
-                           frames: ExecutionFrame[]) {
-  const colon = chalk.gray(':')
+export function printTrace(console: Console, states: QuintEx[],
+                           frames: ExecutionFrame[]): void {
+  const b = chalk.bold
+
   states.forEach((state, index) => {
     assert(state.kind === 'app'
            && state.opcode === 'Rec' && state.args.length % 2 === 0)
 
     if (index < frames.length) {
       // be lenient to broken input
-      out(`[Frame ${index}]`)
-      printExecutionFrameRec(l => out(' ' + l), frames[index], [])
-      out('')
+      console.out(`[${b('Frame ' + index)}]\n`)
+      printExecutionFrameRec(console, frames[index], [])
+      console.out('\n')
     }
 
-    out(`[State ${index}]`)
-    range(0, Math.trunc(state.args.length / 2))
-      .forEach(i => {
-        const key = state.args[2 * i]
-        assert(key.kind === 'str')
-        const valueText = chalkQuintEx(state.args[2 * i + 1])
-        out(` ${key.value}${colon} ${valueText}`)
-      })
-
-    out(''.padStart(80, '─') + '\n')
+    const stateDoc: Doc = [
+      group([brackets(richtext(b, `State ${index}`)), line()]),
+      prettyQuintEx(state)
+    ]
+    console.out(format(console.width, 0, stateDoc))
+    console.out('\n\n')
   })
 }
-