+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 ### Added
+
+- Pretty-printing of expressions in REPL (#870)
+
 ### Changed
 ### Deprecated
 ### Removed